+++ tests/testthat/test-meta-properties.R
@@ -1,12 +1,12 @@
 test_that("fwa_meta_properties works", {
-  rlang::local_options(lifecycle_verbosity = "quiet")
+  rlang::local_options(lifecycle_verbosity = "quiet", nocache = TRUE)
 
   properties <- fwa_meta_properties("whse_basemapping.fwa_stream_networks_sp")
   expect_s3_class(properties, "tbl_df")
 })
 
 test_that("fwa_meta_properties haven't changed for stream networks", {
-  rlang::local_options(lifecycle_verbosity = "quiet")
+  rlang::local_options(lifecycle_verbosity = "quiet", nocache = TRUE)
 
   properties <- fwa_meta_properties("whse_basemapping.fwa_stream_networks_sp")
   expect_snapshot_data(properties, "properties")