+++ NEWS.md
@@ -1,4 +1,21 @@
-<!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
+<!-- NEWS.md is maintained by https://fledge.cynkra.com, contributors should not edit this file -->
+
+# fwapgr 0.1.0.9014
+
+- `downstream_route_measure` of `locate_along_interval()` now returned as an integer together with `index`.
+
+- Indices including blue_line_key, gnis_id and linear_feature_id now returned as integers.
+
+
+# fwapgr 0.1.0.9013
+
+- Added `nocache = getOption("fwa.nocache", FALSE)` argument to all functions.
+
+# fwapgr 0.1.0.9012
+
+- `collection` dropped `id` column.
+- `bc_ind` from `TRUE` to `FALSE` for `index-point` Grostete Creek and two others.
+- In `collection` properties dropped `upstream_area_ha`.
 
 # fwapgr 0.1.0.9011